Moved BaseRepository class into classes folder and updated file paths accordingly
